requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Async and HTTP
aiohttp>=3.9.0
asyncio-mqtt>=0.13.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Utilities
requests>=2.31.0
typing-extensions>=4.5.0

# Optional: Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: Vector databases
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Optional: Monitoring
opentelemetry-distro>=0.40b0
opentelemetry-exporter-jaeger>=1.21.0
